DEV: fixes safari bug with content including linebreaks
Chat had the following bug:

- Paste the following content into chat composer:
```
a

b
```

It would render as (in safari) only:

```
a

b
```

It seems this behavior occurs because of how Safari handles the white-space CSS property in conjunction with the `::placeholder` and `:placeholder-shown` pseudo-classes. Specifically when rendering multiline content in textareas where white-space: nowrap is applied to these pseudo-classes.

I couldn't find an exact source of this, but surely commenting this part of the code was making it work, it seems forcing `pre-wrap` is working.