DEV: allows d-menu to refocus trigger
The trigger will now be refocused after closing the menu by default. This is very useful for tab behavior, otherwise after closing the menu, the body could be focused making it impossible to re-open the menu with Enter or just move up/down the flow.

If you want to disable this behavior:

```javascript
this.menuApi.close({focusTrigger: false});
```

This behavior is already disabled when closing after clicking outside. We don't want to risk interfering with the click of the user.