DEV: Refactor `NotificationSerializer#acting_user_avatar_template`
This commit changes the behavior for how `NotificationSerializer` includes or excludes the `acting_user_avatar_template` attribute. Previously, the inclusion was dependent on the controller preloading the necessary data based on the `show_user_menu_avatars` site setting. This setup is problematic because it blends the concerns of controllers and serializers, introducing tight coupling and redundancy. With this change, the `NotificationSerializer` is solely responsible for how the `acting_user_avatar_template` attribute is included/excluded thus reducing complexity in our controllers.