DEV: Introduce 'dev tools' toolbar and plugin-outlet debugger
This commit introduces a new 'dev tools' feature for core, theme and plugin developers. To enable, run `enableDevTools()` in the browser console. This will load a separate dev-tools JS/CSS bundle, and show a new toolbar on the left of the page. Dev Tools will remain enabled until the 'x' button is clicked, or `disableDevTools()` is run in the console.

The toolbar currently has two buttons:

- "Toggle safe mode" provides an easy way to toggle all themes/plugins on/off

- "Debug plugin outlets" is inspired by the popular 'plugin outlet locations' theme component. It hooks into core's plugin outlet system, and renders a button into every single outlet. Those buttons have a tooltip which shows more information about the outlet, including all of the outletArg values. To inspect the value further, buttons allow the values to be saved to globals and logged to the console.

All of this is implemented under `/static`, and is only async-import()-d when the dev tools are enabled. Therefore, we can continue to add more tools, with zero performance cost to ordinary users of Discourse.