FIX: correctly position emoji picker when clicking more
The emoji picker was correctly showing in both cases, but when clicking we didn't have a valid range as the click makes us lose selection.

This commit uses the coordinates of the mouse on the click event to create a virtual element. We can't create the event target as the menu is disappearing after the click on more.

No tests as we already test it's showing, and testing position is complicated and often unreliable.