DEV: Do not trigger floatkit 'hover' for touch events
When tapping something on a touch-enabled device, `mouseMove` events are still fired, so floatkit would still be triggered even if configured for 'hover' only. For links, this would be particularly strange, because the tooltip would appear for a split-second, before the page navigation occured.

To avoid this problem, we can use the more-modern 'pointerMove' event, and check the `pointerType` to exclude 'touch'