FIX: ensures scrolls work in chat when touch is on text
The stacking context fix we use in chat to avoid: https://bugs.webkit.org/show_bug.cgi?id=262287 was causing this weird behavior in chat where the scroll event wouldn't fire when the finger is on text and not an empty area of the scrollable div.

This simplified implementation seems to work reliably and avoids the issue.