FIX: removes mousewheel edge case on messages
Message action menu used to be much more intrusive and show during scroll, if your mouse was going over a menu during the scroll it would stop it. Now it shouldn't happen anymore and this hack seems un-necessary, moreover, it was causing a bug where it would close the emoji picker unexpectedly if you would scroll a little bit on the actions menu while the picker was opened.