UX: Only lock viewport scale during focusin events for iOS PWA/Hub
c171e3dc works well in Safari, because the browser ignores the `user-scalable=no` directive. However, PWA/Hub do respect the directive, which means that it stopped pinch-zooming from working.

This commit updates the strategy for those environments so that the viewport is only locked briefly during a focusin event. The simpler strategy is maintained for the real safari browser.