PERF: Enqueue `Job::BackfillBadge` in `Jobs::BadgeGrant`
This commit updates the `Jobs::BadgeGrant` scheduled job to enqueue on
`Job::BackfillBadge` regular job for each enabled badge on the site.

The rationale for this change is that we started seeing the
`Jobs::BadgeGrant` job taking hours on sites with lots of enabled badges
as well as users because the job was backfilling all enabled badges
serially within the job. This is bad as it means that a `mini_scheduler` thread is tied up
by this job thus reducing the overall capacity of `mini_scheduler` for
hours.