DEV: Ensure sequential `replaceText` calls work correctly
`replaceText` will replace some text, then call `selectText()` to restore the cursor position. However, selectText was asynchronous, so calling `replaceText()` multiple times in the same runloop iteration would cause the cursor to jump to an unexpected place. This likely explains some of the weird behavior we've seen with upload-markdown replacement