DEV: implements model/afterModel/deactivate in drawer router
These changes allow to load model when loading a screen of the drawer, the underlying idea is to avoid having to rely on the global `activeChannel`, this essentially makes each screen responsible for it's data.

This change is also fixing a bug where clicking on a link routing to the same screen of the drawer you are already on, would display a blank drawer.