DEV: Add proper error response when searching with an invalid page param
Previously, for a search query with `page=11` or higher, we were quietly
returning the page 10 results. The frontend app isn't affected because it
sets its own limit to 10 pages, but still, this response from the search
endpoint does not make sense.

This change switches to returning a 400 error when the `page` parameter
is above the allowed limit (a max of 10).