FIX: hide none on required fields with value
Prior to this fix if a field was marked as required we would never show the none option of the select, however, some forms come with default values and in this case we would show the first choice of the select as selected when it would actually be null. This commit ensures that in this case we show none, and once you select a value, you can't go back to none.