FIX: Remove groups when promotion is recalculated.
The group has `grant_trust_level` setting which automatically updates the trust level when the user is added to the group.

Similarly, when the user is removed from the group, the trust level is recalculated.

There was a bug that when the trust level was downgraded, the user was not removed from automatic groups like for example `trust_level_3`.