DEV: Bump discourse-fonts (#31059)
This brings in the change from ttf to woff2 fonts
for all our fonts.

This brings the total distributed size of our fonts down to 1.4MB (down
from 17MB) which is ~91% decrease, and most fonts are now only around
20KB max individually :slight_smile: The biggest reduction was Noto Sans
JP, which was in .otf format and ~4MB for each file.