DEV: finer format control on field title/description
FormKit allows you to set `@format` on a field which will set the width of the wrapping container of the control. However, title and descriptions are out of this container. FormKit will now by default applies the same format to title and container.

You can override this behavior by using `@titleFormat` and `@descriptionFormat`.