FEATURE: New option for `DMenus` inside composer
This update adds a new option to the `DMenu` options called `insideComposer`. This is to be used when instantiating a new menu inside the composer. Adding this flag will ensure that the menu has an adequate `z-index`, and thus shown to the user without any issues.

This will help each use-case from adding workarounds. For example, in Discourse AI we have been using various workarounds to ensure that the `DMenu` is shown correctly. These can now be safely removed.