DEV: Block running `db:create db:migrate` in single command
By design, `db:create` initializes the Rails app with SKIP_DB=true. That means that SiteSettings get set up with the LocalProcessProvider instead of the DBProvider. In other words: any calls to site settings will return the default, rather then the actual value in the database.

Running `db:migrate` in the same rake invocation means that rails will not be re-initialized, and so skip_db will remain true. Site settings accessed during migrations and fixtures will therefore return incorrect values.

One example of this is that running `bin/rake db:create db:migrate` repeatedly in a development environment will cause the FAQ topic to be seeded repeatedly, because the seed logic does not have access to the site setting which stores the already-seeded topic id.