FIX: better standalone checkbox support
Before this commit it was complicated to render a `Checkbox` outside of a `CheckboxGroup` as you would get no title, no description, no optional hint and not tooltip.

This commits makes all of this possible by adding a special case for checkboxes, and sharing code for tooltips and optional hint.

This commit also uses this opportunity to refactor part of the code to use curryComponent and reduce code duplication.