FIX: do not notify admins about PMs when suppress is on
When `suppress_secured_categories_from_admin` SiteSetting is enabled, it is expected that the admin will not be notified about PMs in which they are not participating - even when they watch the attributed tag.

Before it was only checking if the admin had access to a secured category assigned to a regular topic. PMs do not have categories so we need to ensure that admin in participating in that conversation.