FIX: correctly display label
A regression introduced in 32c8aa0aad880bcab372ffd75bce3c857060d721 incorrectly passes label to the trigger component, but also passes translatedLabel instead of label to the menu.

The existing test was checking for the presence of "label", but it was actually returning true because the test was showing "en.label". The test has been modified in consequences.