FEATURE: Experimental admin search
This feature allows admins to find what they are
looking for in the admin interface via a command
palette. This replaces the admin sidebar filter
as the focus of the Ctrl+/ command, but the sidebar
filter can also still be used. Perhaps at some point
we may remove it or change the shortcut.

The palette presents the following data for filtering:

* The admin nav map, which is also used for the sidebar
* All site settings
* Themes
* Components

Admins can also filter which items are shown in the palette.

This is still **extremely** WIP -- the interface for
the palette is not ideal, it's not keyboard accessible,
the design needs to be refined, the code needs to be
refined, and there are still numerous other considerations
here, like:

* Do we want to include reports?
* We need to include the automatically generated plugin setting
  pages added by Ted
* Do we want to show screenshots for themes and components
  if available, or stick to icons?
* Site setting filters are janky when visitng the same setting
  page with a different filter, the page is not refreshed