DEV: Refactor chat HTML decorating
Uses the new `<DecoratedHtml` component, which takes care of the full decorate/render lifecycle. That means we can drop all the custom modifiers/debouncing which chat was doing. It also naturally adds support for `helper.renderGlimmer` in chat decorations.

This should resolve a number of subtle bugs related to chat message decorations.