DEV: Fix all sticky filters for setting pages
For routes where we have a root config route and we have
an implicit setting route, we need to be able to reset the
site setting filter from the root route. We need to look up
the settings controller explicitly to do this, otherwise inside
resetController we are operating on the parent route which does
not have access to the setting controller which inherits from
AdminAreaSettingsBaseController.

This fixes issues where site setting filters are sticky. We also
apply the same resetController fix manually to actual setting routes
(like the one for Flags) and for the admin site settings route,
and the plugin show settings route which is used by all plugins.