Add PKCS1_OAEP_PADDING option to UserAPIKeySpec
This commit adds a new site setting to the User API Keys options. The
padding option specified here will be used to encrypt the user api key
created and sent back to the client.

The default, PKCS1_PADDING, is the value currently in place (so
existing) integrations won't break),but if a user chooses to use
PKCS1_OAEP_PADDING, they can implement it and use it as they see fit.