DEV: Remove unused 'advanced' theme editor mode
This mode was introduced briefly in 2019, and allowed editing theme settings/translations/extra-css from the admin panel UI. But then we decided that making it possible to edit these fields via the UI conflicted with our recommendation of using a proper IDE, discourse_theme, and version-controlled themes. The feature was disabled in core, and theme-creator uses a monkey-patch to expose it.

Even for the current use in theme creator, there are now major styling and functionality issues, because this feature is not visible or tested in core. For example: it doesn't allow editing 'extra JS' fields, which are arguably the most common.

This commit removes the code for this 'advanced' mode, in advance of some more extensive modernization of this component/model.