DEV: Drop ember-cli-based SCSS and locale compilation
This totally separate SCSS and i18n compilation pipelines only existed so that we could run `ember exam` in CI without starting Rails.

Now that our CI has such heavy caching of Ruby dependencies and database migrations, the speed benefit of this is not worth the cost of maintaining these separate pipelines.

Therefore, this commit removes that system, and updates CI to use `bin/rake qunit:test`. That will start up a Rails server and proxy stylesheet/locale requests to it. This strategy was already used for our theme and plugin qunit test runs.