FIX: Lightbox chat uploads immediately after optimize
When a message is first sent, the original upload URL is used for a split-second before it's replaced by the optimised thumbnail. Using `didInsert` didn't re-run on any changes to the uploads. Switching to a proper modifier, and checking the length of the uploads array within it, will ensure that it is re-run whenever the uploads change.