DEV: Extract 'common' core CSS into own file
Previously we were compiling two CSS bundles: Desktop and Mobile. The majority of both files was the 'common' css. This commit splits those common styles into their own file so that there is less duplication. This should improve compilation times + cache reuse, as well as opening the door for experiments with media-query-based mobile-modes.

The only functional change is that we can no longer use `@extend` to copy common rules to mobile/desktop. This is probably for the best. Duplication and/or mixins are a more native-css pattern for this.