FIX: Drafts should not have an expand button (#31424)
## :mag: Overview

Currently editing a long post and it existing as a draft in the user
stream results in the expand post button being shown in the stream.
However, expanding does nothing and results in errors because it cannot
make the request. This update applies a fix to ensure that long drafts
on edits do not have an expand button

## 📸 Screenshots

### ←Before
![Screenshot 2025-02-20 at 11 14
32](https://github.com/user-attachments/assets/c090d025-6efb-458b-bb6b-4d12e0b214ae)

### →After
![Screenshot 2025-02-20 at 11 14
48](https://github.com/user-attachments/assets/077a1a6a-c6f3-445a-9d65-357acab195f7)