FIX: prevents sidebar to scroll when opening channel (#31454)
This has been a very hard to track and reproduce bug, but after a lot of
profiling I detected that a scroll was showing in the trace just after a
RemoveChild in div.sidebar-sections.

I suspect the performance of this is bad due to a difficulty to
reconciliate object identity and using key seems to help here as I was
not able to reproduce the bug after.