DEV: allows to add a draft without persisting it
On chat setup we get the initial state of drafts for the current users, we need to add them to the drafts manager, but we don't need to store them again on the backend.

This commit adds a persist (boolean) parameter to the `ChatDraftsManager.add` service which when set to true will only add the draft on the frontend and not send it to the backend.

No test, as the behavior is already tested and unchanged, this is only a performance improvement.