DEV: Ensure that Serializers forward their scopes, now and forever
This commit is based on the premise that every serializer should have
a `scope` parameter which is an instance of a `Guardian`.

In many places, serializers which construct other serializers neglect to
forward their `scope` parameter to the sub-serializer.  When one of these
sub-serializers is later modified to make use of its scope in a new way or
under a new circumstance, the missing scope along certain code paths can cause
failures which are hard to debug.  It is difficult to pin down where, within a
chain of serializers, the scope has not been forwarded.

This commit fixes all the current call-sites (in code with test coverage),
where a serializer has neglected to forward its scope.  It also introduces a
mechanism to track where other classes fail to provide a scope when
constructing a serializer.  This mechanism will cause the codebase to evolve
to a state where all serializers are reliably and predictably provided with
a `Guardian` as their scope.

Elements of this commit:

 1. Modify `ApplicationSerializer#new` to log/raise an error if a serializer
    is constructed without a `scope:` argument.  (Error is raised only in a
    test environment.)

 2. Add a `scope: scope` argument to every call-site where a serializer is
    constructed *by another serializer* (forwarding sites) and the `scope:`
    argument is missing.

 3. Define `PlaceholderGuardian`, to be used to provide a `scope:` argument
    wherever one was missing in a non-forwarding serializer construction.

 4. Add a new `scope: PlaceholderGuardian.new` argument to every remaining
    (non-forwarding) call-site where a serializer is constructed without a
    `scope:` argument.  (Call-sites which already have a `scope:` argument
    have been left untouched, even if it assigned the `scope` to `nil`.)