FIX: Sendgrid Webhook payload parsing
When an email bounced because the host does not exist (any more), the payload of the webhook Sendgrid sends does not include a status field, leading to an error: NoMethodError (undefined method []' for nil)`.
Fixed by checking if the field has a value and setting it to an unrecoverable error.