FIX: Include hostname in extra-locales URLs
Followup to 1b5e4b6b0fef9811839490f2ee3b9f31d5fbed3b

When running through a CDN in a multisite environment, the site hostname needs to be included in the path or query or the URL so that the 'current db' can be set correctly by rails multisite. Without this, the response will assume the default site in the cluster.