PERF: Move extra-locale digest from query-param to path
CDNs are often configured to strip query params, which means that the `?v=` parameter wasn't reaching the Rails app, and therefore the cache-control header was not being set correctly. Having a 40 character sha1 digest in the **path** is the approach we take for other similar assets like stylesheets and theme-javascripts.

Also adds a spec for the fix in 573fbeef64f052decc47e740cbe01a3c298c20b5