DEV: replace passwordValidation mixin in CreateAccount modal component (#31444)
This introduces a helper class for password validation logic, and
replaces the mixin in the create-account modal component class. All
properties that impact password validation in that class are also
converted to autotracked properties and getters - the 1 remaining mixin
here (UserFields) doesn't reference properties through `this.model` so
no need to make changes in that mixin.